Add App.Revelator
App.Revelator is a framework library to perform an exception revelator pattern (personal definition)

An exception revelator pattern is a pattern which commonly appears on a programming language which
support try-catch. It use try-catch to escape from deep inside of the call stack.

Vim's try-catch is not powerfull enough to do this nicely, so this module helps to do this a bit
easier.