System.Job: Fix busy-loop issue on Vim 8.1.0349
It seems 'close_cb' is not called from 'sleep 1m' line in '_exit_cb'
callback so that 'a:job.__closed' never become v:true after 8.1.0349.

So check if the channel is open or not and read all buffered messages
in '_exit_cb' instead.

Note that even with this fix, busy-loop MAY occur if '_exit_cb' has
called from 'sleep' in a function in 'statusline' because of 8.1.0342.

Ref: https://github.com/vim-jp/issues/issues/1185