System.Job: Fix behavior in Neovim 0.3.0 and over
It seems Neovim 0.3.0 and over changes the behavior for closing stdout/stderr channel.
It invoke callbacks with [''] when channel has closed while Vim and previous Neovim
silently close the channel.

While the data is a newline splitted string list so receiving [''] should NOT change the
final data structure, I decided to ignore such data in Neovim 0.3.0 and over.

This change is critical for plugins which care about the event itself such as Async.Process
which is under development.